dating-bot/backend/requirements.txt
# Core dependencies
selenium==4.10.0
webdriver-manager==3.8.6
python-dotenv==1.0.0

# For message filtering and analysis
nltk==3.8.1
spacy==3.6.1
scikit-learn==1.3.0
textblob==0.17.1

# Utilities
requests==2.31.0
pandas==2.0.3
numpy==1.24.3

# API dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6